refactor: split module tests, address CodeRabbit review feedback
- Extract ModuleScanFixture to shared header (module_scan_fixture.h)
- Split module import tests into separate file (module_import_tests.cpp)
- Make fixture non-copyable/non-movable to prevent dangling c_str() pointers
- Use exact EXPECT_EQ for partition names instead of substring find()
- Add modules.empty() assertion to NoModule fallback test
